Make GetUsers and GetUser methods asynchronous
Refactor GetUsers method in UsersController to be asynchronous by changing its return type to Task<IActionResult> and using the async keyword. Include user roles in the response by fetching roles with _userManager.GetRolesAsync(user) and adding them to a new userList.

Update GetUser method to include user roles in the response. Fetch roles if a user is found by their ID and include them in the returned object.

Re-indent EnableCors and HttpGet attributes for better readability.